Gen AI Governance Checklist
The document is structured in the form of a Generative AI (Gen AI) Solution Card. It aims to provide the necessary details of the underlying Gen AI solution, such that it can be deployed in a scalable and responsible fashion, complying with applicable regulations. Given this, it serves as an initial governance checklist for the Legal, Privacy, Auditing, Responsible AI, Ethical AI, etc. teams of the enterprise to understand the design, capabilities, and constraints of the Gen AI solution; who can then decide on the further processing steps for their respective streams (as needed). This shows a proactive and responsible approach by the enterprise collaborating with community at large to deploy state-of-the-art technology solutions that benefit society as a whole.